+++ HexaEngine.Vulkan/VulkanGraphicsDevice.cs
@@ -252,8 +252,8 @@
             return new()
             {
                 // TODO: Window size instead of 0
-                Width = (uint)Math.Clamp(0, capabilities.MinImageExtent.Width, capabilities.MaxImageExtent.Width),
-                Height = (uint)Math.Clamp(0, capabilities.MinImageExtent.Height, capabilities.MaxImageExtent.Height)
+                Width = Math.Clamp(0, capabilities.MinImageExtent.Width, capabilities.MaxImageExtent.Width),
+                Height = Math.Clamp(0, capabilities.MinImageExtent.Height, capabilities.MaxImageExtent.Height)
             };
         }
